Feature/wip19 11

+++ src/products/products.service.ts
@@ -4,6 +4,7 @@ import { Repository } from 'typeorm';
 import { Product } from './entities/product.entity';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
+import { RpcException } from '@nestjs/microservices';
 
 @Injectable()
 export class ProductsService {
@@ -22,17 +23,32 @@ export class ProductsService {
   }
 
   async findOne(id: number): Promise<Product> {
-    return this.productsRepository.findOne({ where: { id } });
+    console.log('Finding product with ID:', id); // Log para depurar
+    const product = await this.productsRepository.findOne({ where: { id } });
+    if (!product) {
+      throw new RpcException(`Product with ID ${id} not found`);
+    }
+    console.log('Product found:', product); // Log para verificar
+    return product;
   }
+  
 
   async update(id: number, updateProductDto: UpdateProductDto): Promise<Product> {
     await this.productsRepository.update(id, updateProductDto);
     return this.findOne(id);
   }
 
+  
   async remove(id: number): Promise<void> {
+    const product = await this.productsRepository.findOne({ where: { id } });
+    if (!product) {
+      throw new RpcException(`Product with ID ${id} not found`);
+    }
     await this.productsRepository.delete(id);
   }
+
+  
+  
 
   async searchByName(name?: string) {
     const query = this.productsRepository.createQueryBuilder('product');
